Use int instead of long for opcode encoding

+++ src/main/java/net/consensys/linea/zktracer/opcode/OpCodeData.java
@@ -34,7 +34,7 @@ import net.consensys.linea.zktracer.opcode.stack.StackSettings;
  */
 public record OpCodeData(
     OpCode mnemonic,
-    Long value,
+    int value,
     InstructionFamily instructionFamily,
     StackSettings stackSettings,
     RamSettings ramSettings,
